Plugins: Move getExtensionSettingsPath() to util file
