+++ client/my-sites/plugins/util.js
@@ -1,0 +1,23 @@
+/**
+ * External dependencies
+ */
+import { find, get, includes } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import config from 'config';
+import sectionsModule from 'sections';
+
+export function getExtensionSettingsPath( plugin ) {
+	const pluginSlug = get( plugin, 'slug', '' );
+	const sections = sectionsModule.get();
+	const section = find( sections, ( value => value.name === pluginSlug ) );
+	const env = get( section, 'envId', [] );
+
+	if ( ! includes( env, config( 'env_id' ) ) ) {
+		return;
+	}
+
+	return get( section, 'settings_path' );
+}
